Add repository (#17)
* Add repository

+++ app/api/disease_history.py
@@ -8,6 +8,7 @@ from starlette.responses import FileResponse
 from app.validators.schemes.user_schemes import DiseaseHistoryScheme
 from app.disease_storage.upload_file import FileUploader
 from app.database.disease_history import HistoriesCollection
+from app.api.repository import Repository
 from config import CLEARED_DIR
 
 router = APIRouter()
@@ -20,15 +21,9 @@ async def get_disease_histories(patient_id: str):
     :param patient_id:
     :return: list of histtories
     """
+    histories_patient = Repository.get_histories_by_patient_id(patient_id)
 
-    histories_patient = HistoriesCollection.to_json(HistoriesCollection.get_objs({'patient_id': patient_id},
-                                                                                 fields=('_id', 'title', 'date_updated',
-                                                                                         'status')))
-
-    if not histories_patient:
-        return {'data': {}, 'result': False}, 200
-
-    return {'data': histories_patient, 'result': True}
+    return {'data': histories_patient, 'result': bool(histories_patient)}
 
 
 @router.get('/read_history/{history_id}')
@@ -38,12 +33,9 @@ async def get_history(history_id: str):
     :param history_id: id of history in the database
     :return: histories with history_id
     """
-    history = HistoriesCollection.to_json(HistoriesCollection.get_one_obj({'_id': history_id}))
+    history = Repository.get_history_by_id(history_id)
 
-    if not history['data']:
-        return {'data': {}, 'result': False}, 200
-
-    return history
+    return {'data': history, 'result': bool(history)}
 
 
 @router.post('/disease_history/add')
@@ -53,7 +45,8 @@ async def add_history(history: DiseaseHistoryScheme):
     :param history: dict of data to add to database
     :return:
     """
-    HistoriesCollection.insert_obj(dict(history))
+    Repository.add_history(history)
+
     return {'description': 'Success add', 'result': True}
 
 
@@ -65,11 +58,9 @@ async def update_history(history_id: str, history: DiseaseHistoryScheme):
     :param history: dict of data to add to database
     :return:
     """
-    if history_id in HistoriesCollection.get_ids():
-        HistoriesCollection.update_obj_by_id(history_id, dict(history))
-        return {'description': 'Success update', 'result': True}
-    else:
-        return {'description': 'Can\'n found history by this id', 'result': False}
+    result, status = Repository.update_history(history_id, history)
+
+    return {'description': result, 'result': status}
 
 
 @router.delete('/disease_history/delete/{history_id}')
@@ -79,10 +70,9 @@ async def delete_history(history_id: str):
     :param history_id: id of history in the database
     :return:
     """
-    if history_id in HistoriesCollection.get_ids():
-        HistoriesCollection.delete_obj_by_id(history_id)
-        return {'description': 'Success delete', 'result': True}
-    return {'description': 'Can\'n found history by this id', 'result': False}
+    result, status = Repository.delete_history(history_id)
+
+    return {'description': result, 'result': status}
 
 
 @router.post("/uploadfile/{history_id}")
@@ -94,21 +84,24 @@ async def upload_file(history_id: str, file: UploadFile = File(...)):
     :return:
     """
     if history_id in HistoriesCollection.get_ids():
-        with open(file.filename, "wb") as buffer:
-            shutil.copyfileobj(file.file, buffer)
+        if not Repository.get_history_by_id(history_id).get('file_name'):
+            with open(file.filename, "wb") as buffer:
+                shutil.copyfileobj(file.file, buffer)
 
-        new_filename = str(uuid.uuid4()) + '.' + file.filename.split('.')[-1]
-        os.renames(file.filename, new_filename)
-        shutil.move(new_filename, "app/disease_storage/")
+            new_filename = str(uuid.uuid4()) + '.' + file.filename.split('.')[-1]
+            os.renames(file.filename, new_filename)
+            shutil.move(new_filename, "app/disease_storage/")
 
-        file_uploader = FileUploader()
-        file_uploader.upload_file('app/disease_storage/' + new_filename, new_filename)
+            file_uploader = FileUploader()
+            file_uploader.upload_file('app/disease_storage/' + new_filename, new_filename)
 
-        os.remove('app/disease_storage/' + new_filename)
+            os.remove('app/disease_storage/' + new_filename)
 
-        HistoriesCollection.update_obj_by_id(history_id, {'file_name': new_filename})
+            Repository.update_history(history_id, {'file_name': new_filename})
 
-        return {'description': 'Success add file', 'result': True}
+            return {'description': 'Success add file', 'result': True}
+        else:
+            return {'description': 'File already added', 'result': False}
 
     return {'description': 'Can\'n found history by this id', 'result': False}
 
@@ -116,7 +109,7 @@ async def upload_file(history_id: str, file: UploadFile = File(...)):
 @router.get("/download/{history_id}")
 async def download_file(history_id: str):
     if history_id in HistoriesCollection.get_ids():
-        filename = HistoriesCollection.to_json(HistoriesCollection.get_one_obj({'_id': history_id})['data']['file_name'])
+        filename = Repository.get_history_by_id(history_id).get('file_name')
 
         if filename not in os.listdir(CLEARED_DIR):
             file_uploader = FileUploader()
@@ -124,7 +117,7 @@ async def download_file(history_id: str):
             if filename in file_uploader.list_blobs():
                 file_uploader.download_file(filename)
             else:
-                return {'description': "File not found", 'result': False}
+                return {'description': "File not found", 'result': True}
 
             shutil.move(filename, "app/disease_storage/")
